schoolsList view error solved + duration type for subscription is now 'Long'

+++ school-yard-spring-app/src/main/java/fr/isika/cda/spring/business/service/SubscriptionService.java
@@ -20,7 +20,7 @@ public class SubscriptionService {
 		return (List<Subscription>) subscriptionRepository.findAll();
 	}
 
-	public Subscription createSubscription(String name, Double price, int duration, List<Feature> featureSelected) {
+	public Subscription createSubscription(String name, Double price, Long duration, List<Feature> featureSelected) {
 		Subscription subscription = new Subscription(price, duration, name);
 		subscription.getFeatures().addAll(featureSelected);
 		subscriptionRepository.save(subscription);
